# Optimize  PossibleIntersection and StablePriorityQueue