--- a/tests/PolygonClipper.Tests/PolygonClipper.Tests.csproj
+++ b/tests/PolygonClipper.Tests/PolygonClipper.Tests.csproj
@@ -23,6 +23,7 @@
     <PackageReference Include="xunit" Version="2.4.0" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
     <PackageReference Include="coverlet.collector" Version="1.2.0" />
+    <PackageReference Include="Clipper2" Version="1.5.4" />
   </ItemGroup>
 
   <ItemGroup>