… tests/PolygonClipper.Tests/TestPolygonUtilitiesTests.cs