A rework of #12 and #13 with added unsafe code!

Note: I've discovered that Clipper2 does not return the correct result for the test data so pay attention to our own results only.

CC @stefannikolei 

```
BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.4652)
11th Gen Intel Core i7-11370H 3.30GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK 9.0.302
  [Host]     : .NET 9.0.7 (9.0.725.31616), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 9.0.7 (9.0.725.31616), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
```
Main

| Method   | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|--------- |---------:|---------:|---------:|------:|--------:|-------:|-------:|----------:|------------:|
| Clipper  | 48.65 us | 0.501 us | 0.444 us |  2.90 |    0.03 | 7.8125 | 0.9155 |  47.87 KB |        1.50 |
| Clipper2 | 16.78 us | 0.151 us | 0.118 us |  1.00 |    0.01 | 5.1880 | 0.3357 |  31.84 KB |        1.00 |

PR (PossibleIntersection)

| Method         | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|--------------- |---------:|---------:|---------:|------:|--------:|-------:|-------:|----------:|------------:|
| PolygonClipper | 47.07 us | 0.467 us | 0.414 us |  2.73 |    0.03 | 7.6904 | 0.6714 |  47.23 KB |        1.48 |
| Clipper2       | 17.25 us | 0.160 us | 0.150 us |  1.00 |    0.01 | 5.1880 | 0.3357 |  31.84 KB |        1.00 |

PR (StablePriorityQueue)

| Method         | Mean     | Error    | StdDev   | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|--------------- |---------:|---------:|---------:|------:|--------:|-------:|-------:|----------:|------------:|
| PolygonClipper | 44.23 us | 0.246 us | 0.218 us |  2.60 |    0.02 | 6.7139 | 0.4883 |  41.34 KB |        1.30 |
| Clipper2       | 16.99 us | 0.157 us | 0.139 us |  1.00 |    0.01 | 5.1880 | 0.3357 |  31.84 KB |        1.00 |