@stefannikolei 2025-07-15
LGTM. 

But whoever designed that Unsafe API... .Add to retrieve a value 🙈🙈