@JimBobSquarePants 2025-07-15
> LGTM.
> 
> But whoever designed that Unsafe API... .Add to retrieve a value 🙈🙈

Haha yeah, it can be clunky! 

I just pushed an update for the StablePriorityQueue also. This takes a few percent off the time.